Move connector base class to connectors module
